requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
transformers
torch
scikit-learn
joblib
PyMuPDF
newspaper3k
lxml_html_clean
groq
redis
structlog
slowapi
langdetect
deep-translator
requests
pytest
pytest-asyncio
pytest-cov
python-multipart

